apps/oauth/pyproject.toml
[project]
name = "oauth"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "litestar[jinja]>=2.0",
    "uvicorn>=0.34",
    "httpx>=0.27",
    "attrs>=23.0",
]
